% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_ConVar.R
\name{AF_ConVar}
\alias{AF_ConVar}
\title{Constained variance acquisition function}
\usage{
AF_ConVar(x, fgpi, Cgpi, equal)
}
\arguments{
\item{x}{description}

\item{fgpi}{description}

\item{Cgpi}{description}

\item{equal}{an optional vector containing zeros and ones, whose length equals the number of
constraints, specifying which should be treated as equality constraints (\code{1}) and
which as inequality (\code{0})}
}
\value{
AF
}
\description{
Predictive mean
}
\examples{
B = rbind(c(0, 1), c(0, 1)) 


}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
