% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_EIvsPoF.R
\name{AF_EIvsPoF}
\alias{AF_EIvsPoF}
\title{EI versus PoF}
\usage{
AF_EIvsPoF(x, fgpi, fmean, fsd, Cgpi, fmin)
}
\arguments{
\item{x}{a vector containing a single candidate point}

\item{fgpi}{the GP surrogate model of the objective function}

\item{fmean}{the mean of the objective value}

\item{fsd}{the standard deviation of the objective value}

\item{Cgpi}{the GP surrogate models of the constraints}

\item{fmin}{the best feasible objective value obtained so far}
}
\value{
The AE value(s) at \code{x}.
}
\description{
EI versus PoF
}
\references{
Parr, J. M., A. J. Keane, A. I. Forrester, and C. M. Holden (2012). Infill sampling criteria
for surrogate-based optimization with constraint handling. \emph{Engineering Optimization} 44(10), 1147–1166.
}
\seealso{
\code{\link[EPBO]{AF_ScaledEI}}, \code{\link[EPBO]{AF_EY}}, \code{\link[EPBO]{AF_OOSS}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
