% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_EY.R
\name{AF_EY}
\alias{AF_EY}
\title{Predictive mean acquisition function}
\usage{
AF_EY(x, fgpi, fmean, fsd, Cgpi, rho, equal)
}
\arguments{
\item{x}{description}

\item{fgpi}{description}

\item{fmean}{description}

\item{fsd}{description}

\item{Cgpi}{description}

\item{rho}{description}

\item{equal}{an optional vector containing zeros and ones, whose length equals the number of
constraints, specifying which should be treated as equality constraints (\code{1}) and
which as inequality (\code{0})}
}
\value{
AF
}
\description{
Predictive mean
}
\examples{
B = rbind(c(0, 1), c(0, 1)) 


}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
