% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_EYvsSDY.R
\name{AF_EYvsSDY}
\alias{AF_EYvsSDY}
\title{Predictive standard deviation acquisition function}
\usage{
AF_EYvsSDY(x, fgpi, fmean, fsd, Cgpi, rho, equal)
}
\arguments{
\item{x}{a vector containing a single candidate point; or a \code{matrix} with
multiple candidate points}

\item{fgpi}{the GP surrogate model of the objective function}

\item{fmean}{the mean of the objective value}

\item{fsd}{the standard deviation of the objective value}

\item{Cgpi}{the GP surrogate models of the constraints}

\item{rho}{the penalty parameters}

\item{equal}{an optional vector containing zeros and ones, whose length equals the number of
constraints, specifying which should be treated as equality constraints (\code{1}) and
which as inequality (\code{0})}
}
\value{
The Predictive Standard Deviation at \code{x}.
}
\description{
The predictive standard deviation acquisition function of the PEPBO method
}
\seealso{
\code{\link[EPBO]{AF_ScaledEI}}, \code{\link[EPBO]{AF_OOSS}}, \code{\link[EPBO]{AF_AE}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
