% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_OOSS.R
\name{AF_OOSS}
\alias{AF_OOSS}
\title{Asymmetric Entropy}
\usage{
AF_OOSS(x, fgpi, fnorm, Cgpi, Cnorm)
}
\arguments{
\item{x}{description}

\item{fgpi}{description}

\item{fnorm}{description}

\item{Cgpi}{description}

\item{Cnorm}{description}
}
\value{
AF
}
\description{
Asymmetric Entropy
}
\examples{
B = rbind(c(0, 1), c(0, 1)) 


}
\references{
Pourmohamad, T. and H. K. H. Lee (2022). Bayesian optimization via barrier functions.
\emph{Journal of Computational and Graphical Statistics} 31(1), 74-83.
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
