% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_PEIC.R
\name{AF_PEIC}
\alias{AF_PEIC}
\title{PEIC}
\usage{
AF_PEIC(x, fgpi, Cgpi, fmin, df, point_update)
}
\arguments{
\item{x}{a vector containing a single candidate point}

\item{fgpi}{the GP surrogate model of the objective function}

\item{Cgpi}{the GP surrogate models of the constraints}

\item{fmin}{the best feasible objective value obtained so far}

\item{df}{the lengthscale parameters of the \code{fgpi} surrogate model}

\item{point_update}{the updating points in a single iteration in parallel}
}
\value{
The AE value(s) at \code{x}.
}
\description{
PEIC
}
\references{
Qian, J., Y. Cheng, J. Zhang, J. Liu, and D. Zhan (2021). A parallel constrained efficient global
optimization algorithm for expensive constrained optimization problems. \emph{Engineering Optimization} 53(2), 300–320.
}
\seealso{
\code{\link[EPBO]{AF_ScaledEI}}, \code{\link[EPBO]{AF_EY}}, \code{\link[EPBO]{AF_OOSS}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
