% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_PoF.R
\name{AF_PoF}
\alias{AF_PoF}
\title{PoF}
\usage{
AF_PoF(x, Cgpi, equal, CVthresh = 0)
}
\arguments{
\item{x}{a vector containing a single candidate point}

\item{Cgpi}{the GP surrogate models of the constraints}

\item{equal}{an optional vector containing zeros and ones, whose length equals the number of
constraints, specifying which should be treated as equality constraints (\code{1}) and
which as inequality (\code{0})}

\item{CVthresh}{a threshold used for equality constraints to determine validity for
progress measures; ignored if there are no equality constraints}
}
\value{
The PoF value(s) at \code{x}.
}
\description{
PoF
}
\seealso{
\code{\link[EPBO]{AF_ScaledEI}}, \code{\link[EPBO]{AF_EY}}, \code{\link[EPBO]{AF_OOSS}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
