% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_ScaledEI.R
\name{AF_ScaledEI}
\alias{AF_ScaledEI}
\title{ScaledEI acquisition function}
\usage{
AF_ScaledEI(x, fgpi, fmean, fsd, Cgpi, epbest, rho, equal)
}
\arguments{
\item{x}{a vector containing a single candidate point; or a \code{matrix} with
multiple candidate points}

\item{fgpi}{the GP surrogate model of the objective function}

\item{fmean}{the mean of the objective value}

\item{fsd}{the standard deviation of the objective value}

\item{Cgpi}{the GP surrogate models of the constraints}

\item{epbest}{the best exact penalty value obtained so far}

\item{rho}{the penalty parameters}

\item{equal}{an optional vector containing zeros and ones, whose length equals the number of
constraints, specifying which should be treated as equality constraints (\code{1}) and
which as inequality (\code{0})}
}
\value{
The ScaledEI value(s) at \code{x}.
}
\description{
The Scaled expected improvement acquisition function of the EPBO method
}
\references{
Noe, U. and D. Husmeier (2018). On a new improvement-based acquisition function
for Bayesian optimization. \emph{arXiv:1808.06918}.
}
\seealso{
\code{\link[EPBO]{AF_OOSS}}, \code{\link[EPBO]{AF_EY}}, \code{\link[EPBO]{AF_AE}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
