% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_ScaledEI_MC.R
\name{AF_ScaledEI_MC}
\alias{AF_ScaledEI_MC}
\title{ScaledEI acquisition function}
\usage{
AF_ScaledEI_MC(x, fgpi, fmean, fsd, Cgpi, epbest, rho, equal, N = 1000)
}
\arguments{
\item{x}{description}

\item{fgpi}{description}

\item{fmean}{description}

\item{fsd}{description}

\item{Cgpi}{description}

\item{epbest}{description}

\item{rho}{description}

\item{equal}{an optional vector containing zeros and ones, whose length equals the number of
constraints, specifying which should be treated as equality constraints (\code{1}) and
which as inequality (\code{0})}

\item{N}{positive scalar integer indicating the number of Monte Carlo samples to be used
for calculating ScaledEI}
}
\value{
ScaledEI
}
\description{
Scaled expected improvement
}
\examples{
B = rbind(c(0, 1), c(0, 1)) 


}
\references{
Noe, U. and D. Husmeier (2018). On a new improvement-based acquisition function
for Bayesian optimization. \emph{arXiv:1808.06918}.
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
