% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilog.R
\name{bilog}
\alias{bilog}
\title{(un-)bilog-transform outcomes}
\usage{
bilog(Y)
}
\arguments{
\item{Y}{A training targets}
}
\value{
The (un-)transformed outcome observations.
}
\description{
The Bilog transform is useful for modeling outcome constraints
as it magnifies values near zero and flattens extreme values.
}
\references{
Eriksson, D., & Poloczek, M. (2021, March). Scalable constrained Bayesian optimization.
In \emph{International Conference on Artificial Intelligence and Statistics} (pp. 730-738). PMLR.
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
