% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{min-max normalization}
\usage{
min_max_scale(x, byrow = FALSE)
}
\arguments{
\item{x}{a \code{matrix} or \code{vector}}
}
\value{
The scaled values.
}
\description{
min-max normalization
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
