% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{optim.BM}
\alias{optim.BM}
\title{Barrier Method}
\usage{
optim.BM(
  blackbox,
  B,
  start = 10,
  end = 100,
  Xstart = NULL,
  urate = 10,
  ncandf = function(k) {
     k
 },
  dg.start = c(0.1, 1e-06),
  ab = c(3/2, 8),
  dlim = sqrt(ncol(B)) * c(1/100, 10),
  verb = 2,
  ...
)
}
\arguments{
\item{blackbox}{blackbox of an input (\code{x}), facilitating vectorization on a
\code{matrix} \code{X} thereof,  returning a \code{list}
with elements \code{$obj} containing the (scalar) objective value and \code{$c}
containing a vector of evaluations of the (multiple) constraint function at \code{x}.}

\item{B}{2-column \code{matrix} describing the bounding box.  The number of rows
of the \code{matrix} determines the input dimension (\code{length(x)} in \code{blackbox(x)});
the first column gives lower bounds and the second gives upper bounds}

\item{start}{positive integer giving the number of random starting locations before
sequential design (for optimization) is performed; \code{start >= 6} is
recommended unless \code{Xstart} is non-\code{NULL}; in the current version
the starting locations come from a space-filling design via \code{\link[tgp]{dopt.gp}}}

\item{end}{positive integer giving the total number of evaluations/trials in the
optimization; must have \code{end > start}}

\item{Xstart}{optional matrix of starting design locations in lieu of, or in addition to,
\code{start} random ones;  we recommend \code{nrow(Xstart) + start >= 6}; also must
have \code{ncol(Xstart) = nrow(B)}}

\item{urate}{positive integer indicating  how many optimization trials should pass before
each MLE/MAP update is performed for GP correlation lengthscale
parameter(s)}

\item{ncandf}{function taking a single integer indicating the optimization trial number \code{t}, where
\code{start < t <= end}, and returning the number of search candidates (e.g., for
expected improvement calculations) at round \code{t}; the default setting
allows the number of candidates to grow linearly with \code{t}}

\item{dg.start}{2-vector giving starting values for the lengthscale and nugget parameters
of the GP surrogate model(s) for constraints}

\item{ab}{prior parameters; see \code{\link{darg}} describing the prior used on the
lengthscale parameter during emulation(s) for the constraints}

\item{dlim}{2-vector giving bounds for the lengthscale parameter(s) under MLE/MAP inference,
thereby augmenting the prior specification in \code{ab}}

\item{verb}{a non-negative integer indicating the verbosity level; the larger the value the
more that is printed to the screen}

\item{...}{additional arguments passed to \code{blackbox}}
}
\value{
The output is a \code{list} summarizing the progress of the evaluations of the
blackbox under optimization:
\item{prog }{ vector giving the best feasible (\code{g(x) <= 0 && |h(x)| <= ethresh}) value of the objective over the trials }
\item{xbest }{ vector giving the recommended solution}
\item{obj }{ vector giving the value of the objective for the input under consideration at each trial }
\item{C }{ \code{matrix} giving the value of the constraint function for the input under consideration at each trial}
\item{X }{ \code{matrix} giving the input values at which the blackbox function was evaluated }
}
\description{
Black-box optimization under inequality constraints via an barrier functions.
}
\examples{
# search space
B = rbind(c(-2.25, 2.5), c(-2.25, 1.75))
# Objective and constraint function for use 
MTP = function(x, B=rbind(c(-2.25, 2.5), c(-2.5, 1.75))){
  x = pmin(pmax(B[,1], x), B[,2])
  f = -(cos((x[1]-0.1)*x[2]))^2 - x[1]*sin(3*x[1]+x[2])
  t = atan2(x[1], x[2])
  c = x[1]^2 + x[2]^2 -((2*cos(t)-1/2*cos(2*t)-1/4*cos(3*t)-1/8*cos(4*t))^2) - ((2*sin(t))^2)
  return(list(obj=f, c=c))
}
BM = optim.BM(MTP, B, ncandf = function(k){ 1e3 }, start = 20, end = 120, verb = 0)
# progress, best feasible value of the objective over the trials
BM$prog
# the recommended solution
BM$xbest
}
\references{
Pourmohamad, T. and H. K. H. Lee (2022). Bayesian optimization via barrier functions.
\emph{Journal of Computational and Graphical Statistics} 31(1), 74-83.
}
\seealso{
\code{\link[laGP]{optim.auglag}}, \code{\link[laGP]{optim.efi}}, \code{\link[EPBO]{optim.AE}}, \code{\link[EPBO]{optim.EP}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
\keyword{design}
\keyword{optimize}
