% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIvsPoF.R
\name{optim.EIvsPoF}
\alias{optim.EIvsPoF}
\title{EI versus PoF}
\usage{
optim.EIvsPoF(
  blackbox,
  B,
  nprl = 3,
  start = 10,
  end = 100,
  Xstart = NULL,
  urate = ceiling(10/nprl),
  dg_start = c(0.1 * sqrt(nrow(B)), 1e-06),
  dlim = c(0.001, 10) * sqrt(nrow(B)),
  plotPareto = FALSE,
  verb = 2,
  ...
)
}
\arguments{
\item{blackbox}{blackbox of an input (\code{x}), facilitating vectorization on a
\code{matrix} \code{X} thereof,  returning a \code{list}
with elements \code{$obj} containing the (scalar) objective value and \code{$c}
containing a vector of evaluations of the (multiple) constraint function at \code{x}.}

\item{B}{2-column \code{matrix} describing the bounding box.  The number of rows
of the \code{matrix} determines the input dimension (\code{length(x)} in \code{blackbox(x)});
the first column gives lower bounds and the second gives upper bounds}

\item{nprl}{positive integer giving the number of the parallel points per iteration}

\item{start}{positive integer giving the number of random starting locations before
sequential design (for optimization) is performed; \code{start >= 6} is
recommended unless \code{Xstart} is non-\code{NULL}; in the current version
the starting locations come from a space-filling design via \code{\link[tgp]{dopt.gp}}}

\item{end}{positive integer giving the total number of evaluations/trials in the
optimization; must have \code{end > start}}

\item{Xstart}{optional matrix of starting design locations in lieu of, or in addition to,
\code{start} random ones;  we recommend \code{nrow(Xstart) + start >= 6}; also must
have \code{ncol(Xstart) = nrow(B)}}

\item{urate}{positive integer indicating  how many optimization trials should pass before
each MLE/MAP update is performed for GP correlation lengthscale
parameter(s)}

\item{dlim}{2-vector giving bounds for the lengthscale parameter(s) under MLE/MAP inference}

\item{plotPareto}{\code{logical} indicating if the Pareto plots should be made after each inner iteration;
the plots show two panels tracking the Pareto front, and the Pareto set}

\item{verb}{a non-negative integer indicating the verbosity level; the larger the value the
more that is printed to the screen}

\item{...}{additional arguments passed to \code{blackbox}}

\item{dg.start}{2-vector giving starting values for the lengthscale and nugget parameters
of the GP surrogate model(s) for constraints}
}
\value{
The output is a \code{list} summarizing the progress of the evaluations of the
blackbox under optimization:
\item{prog }{ \code{vector} giving the best feasible (\code{g(x) <= 0 && |h(x)| <= ethresh}) value of the objective over the trials }
\item{xbest }{ \code{vector} giving the recommended solution}
\item{obj }{ \code{vector} giving the value of the objective for the input under consideration at each trial }
\item{C }{ \code{matrix} giving the value of the constraint function for the input under consideration at each trial}
\item{X }{ \code{matrix} giving the input values at which the blackbox function was evaluated }
}
\description{
Black-box optimization under inequality constraints via a bi-objective optimization.
}
\examples{
# search space
B = rbind(c(-2.25, 2.5), c(-2.5, 1.75))
# Objective and constraint function for use 
MTP = function(x){
  f = -(cos((x[1]-0.1)*x[2]))^2 - x[1]*sin(3*x[1]+x[2])
  t = atan2(x[1], x[2])
  c = x[1]^2 + x[2]^2 -((2*cos(t)-1/2*cos(2*t)-1/4*cos(3*t)-1/8*cos(4*t))^2) - ((2*sin(t))^2)
  return(list(obj=f, c=c))
}
EIvsPoF = optim.EIvsPoF(MTP, B, verb = 0)
# progress, best feasible value of the objective over the trials
EIvsPoF$prog
# the recommended solution
EIvsPoF$xbest
}
\references{
Parr, J. M., A. J. Keane, A. I. Forrester, and C. M. Holden (2012). Infill sampling criteria
for surrogate-based optimization with constraint handling. \emph{Engineering Optimization} 44(10), 1147–1166.
}
\seealso{
\code{\link[laGP]{optim.auglag}}, \code{\link[laGP]{optim.efi}}, \code{\link[EPBO]{optim.BM}}, \code{\link[EPBO]{optim.EP}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
\keyword{design}
\keyword{optimize}
