% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbilog.R
\name{unbilog}
\alias{unbilog}
\title{Un-transform bilog-transformed outcomes}
\usage{
unbilog(Y)
}
\arguments{
\item{Y}{A bilog-transfomred targets}
}
\value{
The un-transformed outcome observations.
}
\description{
The Bilog transform is useful for modeling outcome constraints
as it magnifies values near zero and flattens extreme values.
}
\references{
Eriksson, D., & Poloczek, M. (2021, March). Scalable constrained Bayesian optimization.
In \emph{International Conference on Artificial Intelligence and Statistics} (pp. 730-738). PMLR.
}
\seealso{
\code{\link[EPBO]{bilog}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
