% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnormalize.R
\name{unnormalize}
\alias{unnormalize}
\title{Un-normalize min-max function.}
\usage{
unnormalize(X, bounds)
}
\arguments{
\item{X}{\code{matrix}}

\item{bounds}{2-column \code{matrix} describing the bounding box.
the first column gives lower bounds and the second gives upper bounds}
}
\value{
Un-normalized X w.r.t. the provided bounds.
If all elements of \code{X} are contained within \code{[0, 1]^d},
the un-normalized values will be contained within \code{bounds}.
}
\description{
Un-normalizes X w.r.t. the provided bounds.
}
\examples{
X_normalized = matrix(runif(6), ncol=2)
bounds = cbind(c(-1,-1), c(1,1))
X = unnormalize(X_normalized, bounds)
}
\seealso{
\code{\link[EPBO]{normalize}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
