% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_AE.R
\name{AF_AE}
\alias{AF_AE}
\title{Asymmetric Entropy}
\usage{
AF_AE(x, fgpi, fnorm, Cgpi, Cnorm, fmin, alpha1 = 1, alpha2 = 5, omega = 2/3)
}
\arguments{
\item{x}{description}

\item{fgpi}{description}

\item{fnorm}{description}

\item{Cgpi}{description}

\item{Cnorm}{description}

\item{fmin}{description}

\item{alpha1}{description}

\item{alpha2}{description}

\item{omega}{description}
}
\value{
AF
}
\description{
Asymmetric Entropy
}
\examples{
B = rbind(c(0, 1), c(0, 1)) 


}
\references{
Lindberg, D. V. and H. K. Lee (2015). Optimization under constraints by applying an
asymmetric entropy measure. \emph{Journal of Computational and Graphical Statistics} 24(2), 379-393.
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
