% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_OOSS.R
\name{AF_OOSS}
\alias{AF_OOSS}
\title{One over sigma squared (OOSS) acquisition function}
\usage{
AF_OOSS(x, fgpi, fnorm, Cgpi, Cnorm)
}
\arguments{
\item{x}{a vector containing a single candidate point; or a \code{matrix} with
multiple candidate points}

\item{fgpi}{the GP surrogate model of the objective function}

\item{fnorm}{the maximum of the objective}

\item{Cgpi}{the GP surrogate models of the constraints}

\item{Cnorm}{the maxima of the constraints}
}
\value{
The OOSS value(s) at \code{x}.
}
\description{
The one over sigma squared (OOSS) acquisition function of the barrier method
}
\references{
Pourmohamad, T. and H. K. H. Lee (2022). Bayesian optimization via barrier functions.
\emph{Journal of Computational and Graphical Statistics} 31(1), 74-83.
}
\seealso{
\code{\link[EPBO]{AF_ScaledEI}}, \code{\link[EPBO]{AF_EY}}, \code{\link[EPBO]{AF_AE}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
