% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_PoF.R
\name{AF_PoF}
\alias{AF_PoF}
\title{PoF}
\usage{
AF_PoF(x, Cgpi)
}
\arguments{
\item{x}{a vector containing a single candidate point}

\item{Cgpi}{the GP surrogate models of the constraints}
}
\value{
The PoF value(s) at \code{x}.
}
\description{
PoF
}
\seealso{
\code{\link[EPBO]{AF_ScaledEI}}, \code{\link[EPBO]{AF_EY}}, \code{\link[EPBO]{AF_OOSS}}
}
\author{
Jiangyan Zhao \email{zhaojy2017@126.com}
}
